// src/app/models/modelPharmacie/pharmacien.ts
import { User } from 'src/app/models/interfaceUtilisateur/user';

export class Pharmacien implements User{
    dateNaissance: string;
    email: string;
    fcmToken: string;
    genre: any;
    login: string;
    matricule: number;
    nom: string;
    password: string;
    prenoms: string;
    role: number;
    tel: string;
    tel2: string;
    typeEmploye: number;
}